requirements.txt
certifi==2021.10.8
click==8.1.3
dominate==2.6.0
Flask==2.0.3
Flask-Bootstrap==3.3.7.1
Flask-Script==2.0.6
Flask-WTF==1.0.1
gunicorn==20.1.0
importlib-metadata==4.11.3
itsdangerous==2.1.2
Jinja2==3.1.1
KeyGen-Python==0.0.2
MarkupSafe==2.1.1
python-dotenv==0.20.0
visitor==0.1.3
Werkzeug==2.1.2
WTForms==3.0.1
zipp==3.8.0
